requirements.txt
apprise==0.9.9
google_api_python_client==2.66.0
google_auth_oauthlib==0.5.2
